[pigeon] Add an initial example app (#3761)
This creates an initial example app covering Android (Kotlin), iOS and macOS (Swift), and Windows, showing integration of Pigeon into an app directly, without using a plugin. This serves as both a simple runnable example for clients, and as a future source for snippet extraction for documentation. It includes a simple integration test so that CI will ensure that it's actually working.

Since we can't demonstrate Java and Objective-C in the same application, we could consider adding a second example app covering those in the future.

Currently it shows only host API, and only a single trivial method, but we can expand over time as is useful for documentation.

Part of https://github.com/flutter/flutter/issues/66511